Use i16 instead of i32

+++ src-tauri/src/reachability.rs
@@ -10,19 +10,19 @@ use crate::reachability::properties::check_properties;
 
 mod properties;
 
-pub fn create_rg<'a>(marking: Vec<i32>, transition_inputs: Vec<Vec<i32>>, transition_outputs: Vec<Vec<i32>>) -> Result<RGResponse, String> {
+pub fn create_rg<'a>(marking: Vec<i16>, transition_inputs: Vec<Vec<i16>>, transition_outputs: Vec<Vec<i16>>) -> Result<RGResponse, String> {
     let start_time_rg = Instant::now();
     let t = &transition_inputs.len(); // rows
     let p = &transition_inputs.get(0).expect("empty array").len(); // columns
 
-    let t_in: Array2<i32> = vec_vec_to_array2(&transition_inputs, &t, &p);
-    let t_out: Array2<i32> = vec_vec_to_array2(&transition_outputs, &t, &p);
-    let t_effect: Array2<i32> = &t_out - &t_in;
+    let t_in: Array2<i16> = vec_vec_to_array2(&transition_inputs, &t, &p);
+    let t_out: Array2<i16> = vec_vec_to_array2(&transition_outputs, &t, &p);
+    let t_effect: Array2<i16> = &t_out - &t_in;
 
     let state_vec = arr1(&marking);
     let mut queue: Vec<NodeIndex> = Vec::new();
-    let mut graph = DiGraph::<Array1<i32>, i32>::new();
-    let mut all_states_rev: HashMap<Array1<i32>, NodeIndex> = HashMap::new();
+    let mut graph = DiGraph::<Array1<i16>, i16>::new();
+    let mut all_states_rev: HashMap<Array1<i16>, NodeIndex> = HashMap::new();
 
     let mut step_counter = 0;
 
@@ -43,7 +43,7 @@ pub fn create_rg<'a>(marking: Vec<i32>, transition_inputs: Vec<Vec<i32>>, transi
         let mut has_covering = false;
         active.iter()
             .for_each(|&inx| {
-                let new_state: Array1<i32> = fire_transition(&cur_state, &t_effect, inx as usize);
+                let new_state: Array1<i16> = fire_transition(&cur_state, &t_effect, inx as usize);
                 has_covering = has_covering || match insert_next_state(cur_state_idx, new_state, &mut all_states_rev, &mut graph, inx, &mut queue) {
                     None => { false }
                     Some(new_node_index) => { is_covering(&new_node_index, &graph) }
@@ -73,7 +73,7 @@ pub fn create_rg<'a>(marking: Vec<i32>, transition_inputs: Vec<Vec<i32>>, transi
     return Ok(RGResponse::new(graph.node_count(), graph.edge_count(), properties.reversible, properties.liveness, true, "".to_string()));
 }
 
-fn is_covering(new_node: &NodeIndex, graph: &Graph<Array1<i32>, i32>) -> bool {
+fn is_covering(new_node: &NodeIndex, graph: &Graph<Array1<i16>, i16>) -> bool {
     // TODO: check only if transition produces more tokens than it consumes
     // TODO: do a backwards traversal instead
     let new_node_weight = graph.node_weight(*new_node).unwrap();
@@ -96,7 +96,7 @@ fn is_covering(new_node: &NodeIndex, graph: &Graph<Array1<i32>, i32>) -> bool {
     return false
 }
 
-fn is_strictly_greater_than(arr1: &Array1<i32>, arr2: &Array1<i32>) -> bool {
+fn is_strictly_greater_than(arr1: &Array1<i16>, arr2: &Array1<i16>) -> bool {
     let is_greater_than =  arr1.iter().zip(arr2.iter()).all(|(&a, &b)| a >= b) &&
         arr1.iter().zip(arr2.iter()).any(|(&a, &b)| a > b);
 
@@ -105,7 +105,7 @@ fn is_strictly_greater_than(arr1: &Array1<i32>, arr2: &Array1<i32>) -> bool {
     return is_greater_than;
 }
 
-fn insert_next_state(old_state_idx: NodeIndex, new_state: Array1<i32>, all_states_rev: &mut HashMap<Array1<i32>, NodeIndex>, graph: &mut Graph<Array1<i32>, i32>, inx: i32, queue: &mut Vec<NodeIndex>) -> Option<NodeIndex> {
+fn insert_next_state(old_state_idx: NodeIndex, new_state: Array1<i16>, all_states_rev: &mut HashMap<Array1<i16>, NodeIndex>, graph: &mut Graph<Array1<i16>, i16>, inx: i16, queue: &mut Vec<NodeIndex>) -> Option<NodeIndex> {
     let existing = all_states_rev.get(&new_state);
     return match existing {
         None => {
